Update changelog, improve coverage for globular protein unit tests, merge tests, add switch for ideal tests in sample script for globular proteins
