+++ testsuite/CTestTestfile.cmake
@@ -42,7 +42,7 @@ endfunction()
 pymbe_add_test(PATH cph_ideal_tests.py LABELS long THREADS 2)
 pymbe_add_test(PATH grxmc_ideal_tests.py LABELS long THREADS 2)
 pymbe_add_test(PATH gcmc_tests.py LABELS long)
-pymbe_add_test(PATH globular_protein_tests.py LABELS long beyer2024 THREADS 2)
+pymbe_add_test(PATH globular_protein_unit_tests.py LABELS long beyer2024 THREADS 2)
 pymbe_add_test(PATH peptide_tests.py LABELS long beyer2024 THREADS 2)
 pymbe_add_test(PATH weak_polyelectrolyte_dialysis_test.py LABELS long beyer2024)
 
@@ -61,11 +61,10 @@ pymbe_add_test(PATH parameter_test.py)
 pymbe_add_test(PATH henderson_hasselbalch_tests.py)
 pymbe_add_test(PATH calculate_net_charge_unit_test.py)
 pymbe_add_test(PATH setup_salt_ions_unit_tests.py)
-pymbe_add_test(PATH globular_protein_unit_tests.py)
 pymbe_add_test(PATH analysis_tests.py)
 pymbe_add_test(PATH charge_number_map_tests.py)
 pymbe_add_test(PATH generate_coordinates_tests.py)
 pymbe_add_test(PATH reaction_methods_unit_tests.py)
 pymbe_add_test(PATH determine_reservoir_concentrations_unit_test.py)
-pymbe_add_test(PATH read_define_and_create_protein_test.py)
+pymbe_add_test(PATH globular_protein_unit_tests.py)
 pymbe_add_test(PATH lattice_builder.py)
